lsm-optimization/requirements.txt
numpy
pandas
deap
optuna
matplotlib
requests
sagemaker-inference
flask
boto3
python-dotenv